Convert es5 function syntax to es6 arrow syntax
uncomment the entire code and change the ES5 anonymous functions assigned to
the variables; myFunction, anotherFunction, add, subtract and the one used within
the exampleArray.map() method to ES6 arrow functions.

prevent the exampleArray variable from being global and mutable, by making it
a constant variable with the `const` keyword.