Manage car dealer's inventory
Look up the car with id: 33 and log it's year, make and model to the
console. Get and log the make and model of the last car in the
inventory to the console.Get and sort the list of the car models in
alphabetical order, log the result to the console.

Create a new list from the inventory containing only the car years and
log the newly created list to the console. Create a new list of cars
made before the year 2000 from the `car years list` and log it's length
to console. Create a new list off the inventory containing BMW and Audi
cars, use JSON.stringify() to log the result to the console.